Revert "Project renamed to TriviaApp, delete data_co_checker not web compatible"
