+++ lib/core/usecases/usecase.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
-import 'package:triviaapp/core/error/failures.dart';
+import 'package:rogueteam/core/error/failures.dart';
 
 abstract class UseCase<Type, Params> {
   Future<Either<Failure, Type>> call(Params params);
