+++ lib/core/util/input_converter.dart
@@ -1,5 +1,5 @@
 import 'package:dartz/dartz.dart';
-import 'package:triviaapp/core/error/failures.dart';
+import 'package:rogueteam/core/error/failures.dart';
 
 class InputConverter {
   Either<Failure, int> stringToUnsignedInteger(String str) {
