+++ lib/features/number_trivia/data/models/number_trivia_model.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/cupertino.dart';
-import 'package:triviaapp/features/number_trivia/domain/entities/number_trivia.dart';
+import 'package:rogueteam/features/number_trivia/domain/entities/number_trivia.dart';
 import 'package:meta/meta.dart';
 
 class NumberTriviaModel extends NumberTrivia {
