+++ lib/features/number_trivia/domain/repositories/number_trivia_repository.dart
@@ -1,6 +1,6 @@
 import 'package:dartz/dartz.dart';
-import 'package:triviaapp/core/error/failures.dart';
-import 'package:triviaapp/features/number_trivia/domain/entities/number_trivia.dart';
+import 'package:rogueteam/core/error/failures.dart';
+import 'package:rogueteam/features/number_trivia/domain/entities/number_trivia.dart';
 
 abstract class NumberTriviaRepository {
   Future<Either<Failure, NumberTrivia>> getConcreteNumberTrivia(int number);
