+++ lib/features/number_trivia/domain/usecases/get_concrete_number_trivia.dart
@@ -1,9 +1,9 @@
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
-import 'package:triviaapp/core/error/failures.dart';
-import 'package:triviaapp/core/usecases/usecase.dart';
-import 'package:triviaapp/features/number_trivia/domain/entities/number_trivia.dart';
-import 'package:triviaapp/features/number_trivia/domain/repositories/number_trivia_repository.dart';
+import 'package:rogueteam/core/error/failures.dart';
+import 'package:rogueteam/core/usecases/usecase.dart';
+import 'package:rogueteam/features/number_trivia/domain/entities/number_trivia.dart';
+import 'package:rogueteam/features/number_trivia/domain/repositories/number_trivia_repository.dart';
 import 'package:meta/meta.dart';
 
 class GetConcreteNumberTrivia implements UseCase<NumberTrivia, Params> {
