+++ lib/features/number_trivia/domain/usecases/get_random_number_trivia.dart
@@ -1,8 +1,8 @@
-import 'package:triviaapp/core/error/failures.dart';
+import 'package:rogueteam/core/error/failures.dart';
 import 'package:dartz/dartz.dart';
-import 'package:triviaapp/core/usecases/usecase.dart';
-import 'package:triviaapp/features/number_trivia/domain/entities/number_trivia.dart';
-import 'package:triviaapp/features/number_trivia/domain/repositories/number_trivia_repository.dart';
+import 'package:rogueteam/core/usecases/usecase.dart';
+import 'package:rogueteam/features/number_trivia/domain/entities/number_trivia.dart';
+import 'package:rogueteam/features/number_trivia/domain/repositories/number_trivia_repository.dart';
 
 class GetRandomNumberTrivia implements UseCase<NumberTrivia, NoParams> {
   final NumberTriviaRepository repository;
