+++ lib/features/number_trivia/presentation/bloc/number_trivia_bloc.dart
@@ -4,12 +4,12 @@ import 'package:bloc/bloc.dart';
 import 'package:dartz/dartz.dart';
 import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
-import 'package:triviaapp/core/error/failures.dart';
-import 'package:triviaapp/core/usecases/usecase.dart';
-import 'package:triviaapp/core/util/input_converter.dart';
-import 'package:triviaapp/features/number_trivia/domain/entities/number_trivia.dart';
-import 'package:triviaapp/features/number_trivia/domain/usecases/get_concrete_number_trivia.dart';
-import 'package:triviaapp/features/number_trivia/domain/usecases/get_random_number_trivia.dart';
+import 'package:rogueteam/core/error/failures.dart';
+import 'package:rogueteam/core/usecases/usecase.dart';
+import 'package:rogueteam/core/util/input_converter.dart';
+import 'package:rogueteam/features/number_trivia/domain/entities/number_trivia.dart';
+import 'package:rogueteam/features/number_trivia/domain/usecases/get_concrete_number_trivia.dart';
+import 'package:rogueteam/features/number_trivia/domain/usecases/get_random_number_trivia.dart';
 
 part 'number_trivia_event.dart';
 part 'number_trivia_state.dart';
