+++ lib/injection_container.dart
@@ -1,3 +1,4 @@
+import 'package:data_connection_checker/data_connection_checker.dart';
 import 'package:get_it/get_it.dart';
 import 'package:http/http.dart' as http;
 import 'package:shared_preferences/shared_preferences.dart';
@@ -55,4 +56,5 @@ Future<void> init() async {
 /*   final sharedPreferences = await SharedPreferences.getInstance(); */
   sl.registerLazySingletonAsync(() => SharedPreferences.getInstance());
   sl.registerLazySingleton(() => http.Client());
+  sl.registerLazySingleton(() => DataConnectionChecker());
 }
