+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: triviaapp
+name: rogueteam
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
@@ -26,6 +26,7 @@ dependencies:
   connectivity: ^3.0.3
   cupertino_icons: ^1.0.2
   dartz: ^0.9.2
+  data_connection_checker: ^0.3.4
   equatable: ^2.0.0
   flutter_bloc: ^7.0.0
   get_it: ^6.0.0
