+++ test/features/number_trivia/data/datasources/number_trivia_remote_datasource_test.dart
@@ -2,9 +2,9 @@ import 'dart:convert';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
-import 'package:triviaapp/core/error/exceptions.dart';
-import 'package:triviaapp/features/number_trivia/data/datasources/number_trivia_remote_datasource.dart';
-import 'package:triviaapp/features/number_trivia/data/models/number_trivia_model.dart';
+import 'package:rogueteam/core/error/exceptions.dart';
+import 'package:rogueteam/features/number_trivia/data/datasources/number_trivia_remote_datasource.dart';
+import 'package:rogueteam/features/number_trivia/data/models/number_trivia_model.dart';
 import 'package:http/http.dart' as http;
 
 import '../../../../fixtures/fixture_reader.dart';
