+++ test/features/number_trivia/domain/usecases/get_random_number_trivia_test.dart
@@ -1,10 +1,10 @@
 import 'package:dartz/dartz.dart';
 import 'package:mockito/mockito.dart';
-import 'package:triviaapp/core/usecases/usecase.dart';
-import 'package:triviaapp/features/number_trivia/domain/entities/number_trivia.dart';
-import 'package:triviaapp/features/number_trivia/domain/repositories/number_trivia_repository.dart';
+import 'package:rogueteam/core/usecases/usecase.dart';
+import 'package:rogueteam/features/number_trivia/domain/entities/number_trivia.dart';
+import 'package:rogueteam/features/number_trivia/domain/repositories/number_trivia_repository.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:triviaapp/features/number_trivia/domain/usecases/get_random_number_trivia.dart';
+import 'package:rogueteam/features/number_trivia/domain/usecases/get_random_number_trivia.dart';
 
 class MockNumberTriviaRepository extends Mock
     implements NumberTriviaRepository {}
