+++ test/features/number_trivia/presentation/bloc/number_trivia_bloc_test.dart
@@ -1,13 +1,13 @@
 import 'package:dartz/dartz.dart';
 import 'package:mockito/mockito.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:triviaapp/core/error/failures.dart';
-import 'package:triviaapp/core/usecases/usecase.dart';
-import 'package:triviaapp/core/util/input_converter.dart';
-import 'package:triviaapp/features/number_trivia/domain/entities/number_trivia.dart';
-import 'package:triviaapp/features/number_trivia/domain/usecases/get_concrete_number_trivia.dart';
-import 'package:triviaapp/features/number_trivia/domain/usecases/get_random_number_trivia.dart';
-import 'package:triviaapp/features/number_trivia/presentation/bloc/number_trivia_bloc.dart';
+import 'package:rogueteam/core/error/failures.dart';
+import 'package:rogueteam/core/usecases/usecase.dart';
+import 'package:rogueteam/core/util/input_converter.dart';
+import 'package:rogueteam/features/number_trivia/domain/entities/number_trivia.dart';
+import 'package:rogueteam/features/number_trivia/domain/usecases/get_concrete_number_trivia.dart';
+import 'package:rogueteam/features/number_trivia/domain/usecases/get_random_number_trivia.dart';
+import 'package:rogueteam/features/number_trivia/presentation/bloc/number_trivia_bloc.dart';
 import 'package:bloc_test/bloc_test.dart';
 
 class MockGetConcreteNumberTrivia extends Mock
