+++ test/widget_test.dart
@@ -8,7 +8,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:triviaapp/main.dart';
+import 'package:rogueteam/main.dart';
 
 void main() {
 /*   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
