+++ web/index.html
@@ -20,13 +20,13 @@
   <!-- iOS meta tags & icons -->
   <meta name="apple-mobile-web-app-capable" content="yes">
   <meta name="apple-mobile-web-app-status-bar-style" content="black">
-  <meta name="apple-mobile-web-app-title" content="triviaapp">
+  <meta name="apple-mobile-web-app-title" content="rogueteam">
   <link rel="apple-touch-icon" href="icons/Icon-192.png">
 
   <!-- Favicon -->
   <link rel="icon" type="image/png" href="favicon.png"/>
 
-  <title>triviaapp</title>
+  <title>rogueteam</title>
   <link rel="manifest" href="manifest.json">
 </head>
 <body>
