+++ web/manifest.json
@@ -1,6 +1,6 @@
 {
-    "name": "triviaapp",
-    "short_name": "triviaapp",
+    "name": "rogueteam",
+    "short_name": "rogueteam",
     "start_url": ".",
     "display": "standalone",
     "background_color": "#0175C2",
